+++ inst/tinytest/test_breaks.R
@@ -25,3 +25,4 @@ expect_equal(mf_get_breaks(x = x, breaks = c(1,10,20,100), k = 1),
              c(1,10,20,100))
 expect_equal(mf_get_breaks(x = x, breaks = "fisher"),
              c(11929, 13953, 15685.5, 17372, 18622, 20354.5, 21761))
+expect_error(mf_get_breaks(x = c(0,x), breaks = "geom"))