+++ NEWS.md
@@ -7,7 +7,7 @@
 - add a test for "x" class in mf_map() 
 - enable proportional symbol plots for single points (type %in% c("prop", "prop_typo", "prop_choro"); see #45
 - fix a bug in color assignment when there is a mismatch between the number of classes and the size of the color palette
-
+- add an error for method "geom" in mf_get_breaks() when min(x) <= 0
 
 ## Feat
 - add "prop_typo"" maps for LINES objects